src/requirements.txt
python-kasa[shell]
